Move retrieve_host_list to dbcopy client
